Server/requirements.txt
django 
djangorestframework 
mysqlclient 
drf-yasg 
gunicorn